discord: Move the notifier into new class
