+++ discord/index.ts
@@ -1,20 +1,15 @@
-import {VoiceConnectionStatus, joinVoiceChannel} from '@discordjs/voice';
+import {joinVoiceChannel} from '@discordjs/voice';
 import type {DiscordGatewayAdapterCreator} from '@discordjs/voice';
-import type {ContextBlock, WebAPICallResult} from '@slack/web-api';
-import {Mutex} from 'async-mutex';
-import Discord, {TextChannel, Collection, Snowflake, GuildMember, VoiceChannel, GatewayIntentBits} from 'discord.js';
+import Discord, {TextChannel, VoiceChannel, GatewayIntentBits} from 'discord.js';
 import logger from '../lib/logger';
 import type {SlackInterface} from '../lib/slack';
 import {getMemberIcon, getMemberName} from '../lib/slackUtils';
 import State from '../lib/state';
 import Hayaoshi from './hayaoshi';
+import {Notifier} from './notifier';
 import TTS from './tts';
 
 const log = logger.child({bot: 'discord'});
-
-interface ChatPostMessageResult extends WebAPICallResult {
-	ts: string;
-}
 
 interface StateObj {
 	users: {discord: string, slack: string}[],
@@ -33,38 +28,13 @@ const discord = new Discord.Client({
 
 discord.login(process.env.TSGBOT_DISCORD_TOKEN);
 
-const mutex = new Mutex();
-
-const getMembersBlock = (roomName: string, members: Collection<Snowflake, GuildMember>) => (
-	{
-		type: 'context',
-		elements: [
-			{
-				type: 'mrkdwn',
-				text: `*[${roomName}]*`,
-			},
-			...Array.from(members.values()).slice(0, 8)
-				.map((member) => (
-					{
-						type: 'image',
-						image_url: member.user.displayAvatarURL({extension: 'png', size: 64}),
-						alt_text: member.displayName,
-					}
-				)),
-			{
-				type: 'plain_text',
-				emoji: true,
-				text: `${members.size} users`,
-			},
-		],
-	} as ContextBlock
-);
-
 export default async ({webClient: slack, eventClient}: SlackInterface) => {
 	const state = await State.init<StateObj>('discord', {
 		users: [],
 		ttsDictionary: [{key: 'https?:\\S*', value: 'URL省略'}],
 	});
+
+	const notifier = new Notifier(slack);
 
 	const joinVoiceChannelFn = (channelId: string = process.env.DISCORD_SANDBOX_VOICE_CHANNEL_ID) => {
 		const channel = discord.channels.cache.get(channelId) as VoiceChannel;
@@ -75,20 +45,6 @@ export default async ({webClient: slack, eventClient}: SlackInterface) => {
 			adapterCreator: channel.guild.voiceAdapterCreator as unknown as DiscordGatewayAdapterCreator,
 		});
 		return connection;
-	};
-	const roomNotifyCache = {
-		lastUnixTime: 0, usernames: <string[]>[], ts: '', action: '',
-	};
-	const notifyCacheLimit = 60000; // 1min
-	const usernameSummarizer = (usernames: string[]) => {
-		if (usernames.length >= 3) {
-			return `＊${usernames[0]}＊, ＊${usernames[1]}＊, ほか${usernames.length - 2}名`;
-		} else if (usernames.length === 2) {
-			return `＊${usernames[0]}＊, ＊${usernames[1]}＊`;
-		} else if (usernames.length === 1) {
-			return `＊${usernames[0]}＊`;
-		}
-		return '';
 	};
 
 	let hayaoshi = new Hayaoshi(joinVoiceChannelFn, state.users);
@@ -131,159 +87,8 @@ export default async ({webClient: slack, eventClient}: SlackInterface) => {
 		tts.onMessage(message);
 	});
 
-	const postMessage = async (
-		{text, count, rooms, ts}: {
-			text: string,
-			count: number,
-			rooms: {
-				name: string,
-				members: Discord.Collection<string, Discord.GuildMember>,
-			}[],
-			ts: string,
-		},
-	) => {
-		const countText = count === null ? '' : `現在のアクティブ人数 ${count}人`;
-
-		if (ts) {
-			const result = await slack.chat.update({
-				ts,
-				channel: process.env.CHANNEL_SANDBOX,
-				username: 'Discord',
-				icon_emoji: ':discord:',
-				text: `${text}\n${countText}`,
-				blocks: [
-					{
-						type: 'section',
-						text: {
-							type: 'mrkdwn',
-							text,
-						},
-					},
-					...rooms.map((room) => getMembersBlock(room.name, room.members)),
-				],
-			}) as ChatPostMessageResult;
-			return result;
-		}
-
-		const response = await slack.chat.postMessage({
-			channel: process.env.CHANNEL_SANDBOX,
-			username: 'Discord',
-			icon_emoji: ':discord:',
-			text: `${text}\n${countText}`,
-			blocks: [
-				{
-					type: 'section',
-					text: {
-						type: 'mrkdwn',
-						text,
-					},
-				},
-				...rooms.map((room) => getMembersBlock(room.name, room.members)),
-			],
-		}) as ChatPostMessageResult;
-		return response;
-	};
-
 	discord.on('voiceStateUpdate', (oldState, newState) => {
-		if (oldState.member.user.bot) {
-			return;
-		}
-
-		const username = oldState.member.displayName;
-		const eventTime = Date.now();
-
-		mutex.runExclusive(async () => {
-			// leave
-			if (oldState.channel !== null && newState.channel === null) {
-				const roomName = oldState.channel.name;
-				const roomId = oldState.channel.id;
-				const count = oldState.channel.members.size;
-				const actionName = `leave-${roomId}`;
-				const update = roomNotifyCache.lastUnixTime + notifyCacheLimit > eventTime && roomNotifyCache.action === actionName;
-
-				if (update) {
-					roomNotifyCache.usernames.push(username);
-				} else {
-					roomNotifyCache.usernames = [username];
-				}
-
-				const response = await postMessage({
-					text: `${usernameSummarizer(roomNotifyCache.usernames)}が<https://discord.com/channels/${process.env.DISCORD_SERVER_ID}/${roomId}|${roomName}>からログアウトしました`,
-					count,
-					rooms: [{name: roomName, members: oldState.channel.members}],
-					ts: update ? roomNotifyCache.ts : null,
-				});
-
-				if (!update && response.ok) {
-					roomNotifyCache.ts = response.ts;
-				}
-
-				roomNotifyCache.action = actionName;
-				roomNotifyCache.lastUnixTime = eventTime;
-			}
-
-			// join
-			if (newState.channel !== null && oldState.channel === null) {
-				const roomName = newState.channel.name;
-				const roomId = newState.channel.id;
-				const count = newState.channel.members.size;
-				const actionName = `join-${roomId}`;
-				const update = roomNotifyCache.lastUnixTime + notifyCacheLimit > eventTime && roomNotifyCache.action === actionName;
-
-				if (update) {
-					roomNotifyCache.usernames.push(username);
-				} else {
-					roomNotifyCache.usernames = [username];
-				}
-
-				const response = await postMessage({
-					text: `${usernameSummarizer(roomNotifyCache.usernames)}が<https://discord.com/channels/${process.env.DISCORD_SERVER_ID}/${roomId}|${roomName}>にログインしました`,
-					count,
-					rooms: [{name: roomName, members: newState.channel.members}],
-					ts: update ? roomNotifyCache.ts : null,
-				});
-
-				if (!update && response.ok) {
-					roomNotifyCache.ts = response.ts;
-				}
-
-				roomNotifyCache.action = actionName;
-				roomNotifyCache.lastUnixTime = eventTime;
-			}
-
-			// move
-			if (oldState.channel !== null && newState.channel !== null && oldState.channel.id !== newState.channel.id) {
-				const newRoomName = newState.channel.name;
-				const newRoomId = newState.channel.id;
-				const oldRoomName = oldState.channel.name;
-				const oldRoomId = oldState.channel.id;
-				const actionName = `join-${oldRoomId}-${newRoomId}`;
-				const update = roomNotifyCache.lastUnixTime + notifyCacheLimit > eventTime && roomNotifyCache.action === actionName;
-
-				if (update) {
-					roomNotifyCache.usernames.push(username);
-				} else {
-					roomNotifyCache.usernames = [username];
-				}
-
-				const response = await postMessage({
-					text: `${usernameSummarizer(roomNotifyCache.usernames)}が<https://discord.com/channels/${process.env.DISCORD_SERVER_ID}/${oldRoomId}|${oldRoomName}>から<https://discord.com/channels/${process.env.DISCORD_SERVER_ID}/${newRoomId}|${newRoomName}>に移動しました`,
-					count: null,
-					rooms: [
-						{name: oldRoomName, members: oldState.channel.members},
-						{name: newRoomName, members: newState.channel.members},
-					],
-					ts: update ? roomNotifyCache.ts : null,
-				});
-
-				if (!update && response.ok) {
-					roomNotifyCache.ts = response.ts;
-				}
-
-				roomNotifyCache.action = actionName;
-				roomNotifyCache.lastUnixTime = eventTime;
-			}
-		});
+		notifier.voiceStateUpdate(oldState, newState);
 	});
 
 	eventClient.on('message', async (message) => {
